minor - refactoring in favor of poly_commit lib export
